Switch from lru_time_cache to moka cache in access token auth client (#51977)
Moka implements TTL (rather than TTI).

From moka docs

```
Cache supports the following cache-level expiration policies:

Time to live (TTL): A cached entry will be expired after the specified
duration past from insert.
Time to idle (TTI): A cached entry will be expired after the specified
duration past from get or insert.
```

The lru_time_cache implemented a time-to-idle and called it
time-to-live (which I confirmed by reading the source).

For the security of access token revocation, we need to make sure the
TTL is actually a TTL, so that it's guaranteed to expire and revalidate,
even on popular accounts that make authed admin calls often.

I looked briefly into layering MokaManager into the HttpClient, but because the HTTP Client is shared with other callers, and the is_authorized_for_deployment method is a POST, this would be nontrivial, so I went with this more-point-fix.

GitOrigin-RevId: 3c22f0383976625cc76c983e6878ed1b401e91d4

+++ crates/common/src/knobs.rs
@@ -1419,7 +1419,7 @@ pub static REQUEST_TRACE_SAMPLE_CONFIG: LazyLock<SamplingConfig> = LazyLock::new
 });
 
 /// Size of the cache for access token authentication
-pub static AUTH_CACHE_SIZE: LazyLock<usize> = LazyLock::new(|| env_config("AUTH_CACHE_SIZE", 1000));
+pub static AUTH_CACHE_SIZE: LazyLock<u64> = LazyLock::new(|| env_config("AUTH_CACHE_SIZE", 1000));
 
 /// Length of time an entry to the access authentication cache is valid
 pub static AUTH_CACHE_TTL_SECONDS: LazyLock<u64> =
